requirements.txt
fastapi
uvicorn[standard]
celery
redis
pydantic
pydantic-settings
requests
python-dotenv
pypdf
beautifulsoup4
google-genai
python-multipart
